% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpphnsw-package.R
\docType{package}
\name{RcppHnsw-package}
\alias{RcppHnsw-package}
\alias{HnswL2}
\alias{Rcpp_HnswL2-class}
\title{Rcpp bindings for the HNSW C++ library for approximate nearest neighbors.}
\description{
HNSW is a library implementing the Hierarchical Navigable Small World method
for approximate nearest neighbor search.
}
\details{
Details about Annoy are available at the reference listed below.
}
\references{
\url{https://github.com/nmslib/hnsw}
}
\author{
James Melville for the R inteface; Yury Malkov for HNSW itself.

Maintainer: James Melville <jlmelville@gmail.com>
}
