% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpphnsw-package.R
\docType{package}
\name{RcppHnsw-package}
\alias{RcppHnsw-package}
\alias{HnswL2}
\alias{Rcpp_HnswL2-class}
\alias{HnswCosine}
\alias{Rcpp_HnswCosine-class}
\alias{HnswIp}
\alias{Rcpp_HnswIp-class}
\title{Rcpp bindings for the hnswlib C++ library for approximate nearest neighbors.}
\description{
hnswlib is a library implementing the Hierarchical Navigable Small World
method for approximate nearest neighbor search.
}
\details{
Details about hnswlib are available at the reference listed below.
}
\references{
\url{https://github.com/nmslib/hnswlib}

Malkov, Y. A., & Yashunin, D. A. (2016).
Efficient and robust approximate nearest neighbor search using Hierarchical
Navigable Small World graphs.
\emph{arXiv preprint} \emph{arXiv:1603.09320}.
}
\author{
James Melville for the R interface; Yury Malkov for hnswlib itself.

Maintainer: James Melville \href{mailto:jlmelville@gmail.com}{jlmelville@gmail.com}
}
